Add i18n support to Credits (#20)
We change the class component to a function component as we have no
state to take care of. Also, we can use the i18n hook for our
translations.
We use the useEffect hook instead of the componentDidMount
lifecycle method as with the empty array it has the same effect.